feat: move authFetch test and Session test to dataunion directory (effectively disabling the tests)
